GA-Based Context Evolution Prototype/pyproject.toml
[project]
name = "ga-context-evolution"
version = "0.1.0"
description = "Prototype for Genetic Algorithm based Context Evolution using LangGraph"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "loguru>=0.7.3",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
